requirements.txt
websockets>=12.0
langchain>=0.3.27
langchain-openai>=0.3.35
langgraph>=0.6.11
bleak==0.22.0
python-escpos>=3.0
pyusb>=1.0.0
faster-whisper
numpy
sounddevice
piper-tts
SpeechRecognition
pyaudio